Specs for database migrations should run in own DB
We have a test that tests the installation and removal of database
extensions.

We have a test that tests a customer can *use* the extensions helper
in their own migrations.

These need to be isolated from the rest of the specs that depend on a
global set of applied migrations.