WIP not-yet-functioning uniqueness validator
I think it's matching on the literal column value instead of decrypting
the value first.

Why moving to ActiveRecordExtensions? Because keeping it in
`Protect::Model::...` results in ActiveRecord looking for its own
`UniquenessValidator` in `Protect::Model` for... reasons. Sufficed to
say, you need to put the validator in a different namespace.